6 1 add test

See merge request wb/ozmitel-daria!12

+++ app/src/main/java/com/example/wbproject/ui/theme/screens/community/CommunityListScreen.kt
@@ -21,7 +21,7 @@ import org.koin.androidx.compose.koinViewModel
 @Composable
 fun CommunityListScreen(
     modifier: Modifier = Modifier,
-    onCommunityCardClickListener: () -> Unit = {}
+    onCommunityCardClickListener: (Int) -> Unit = {}
 ) {
     var searchText by remember {
         mutableStateOf("")
